+++ test/support/net.rb
@@ -27,3 +27,26 @@ class TestNet < Torch::NN::Module
     num_features
   end
 end
+
+class SimpleResidualBlock < Torch::NN::Module
+  def initialize
+    super
+
+    @relu = Torch::NN::ReLU.new
+
+    @seq = Torch::NN::Sequential.new(
+      Torch::NN::Conv2d.new(64, 128, 3, padding: 1, bias: false),
+      Torch::NN::BatchNorm2d.new(128),
+      @relu,
+      Torch::NN::Conv2d.new(128, 128, 3, padding: 1, bias: false),
+      Torch::NN::BatchNorm2d.new(128),
+      @relu,
+      Torch::NN::Conv2d.new(128, 64, 3, bias: false),
+      Torch::NN::BatchNorm2d.new(64)
+    )
+  end
+
+  def forward(x)
+    @relu.forward(@seq.forward(x) + x)
+  end
+end